# server/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "MTC API Documentation",
    "version": "1.0.0",
    "title": "MTC API Documentation"
  },
  "host": "localhost:10010",
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "parameters": {
    "clientId": {
      "name": "clientId",
      "type": "string",
      "in": "header",
      "required": true,
      "description": "APIs are secured by clientId."
    },
    "clientSecret": {
      "name": "clientSecret",
      "type": "string",
      "in": "header",
      "required": true,
      "description": "clientSecret"
    },
    "guest": {
      "name": "guest",
      "type": "string",
      "in": "header",
      "required": true,
      "description": "guest id"
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/open/offerImage": {
      "parameters": [
        {
          "$ref": "#/parameters/clientId"
        },
        {
          "$ref": "#/parameters/clientSecret"
        },
        {
          "$ref": "#/parameters/guest"
        }
      ],
      "get": {
        "tags": [
          "OfferImage"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for getOfferImage",
        "consumes": [
          "application/json"
        ],
        "operationId": "getOfferImage",
        "description": "API for get Cart\n",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "minimum": 0,
            "required": false,
            "type": "number"
          },
          {
            "in": "query",
            "name": "id",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "item",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "limit",
            "minimum": 0,
            "maximum": 20,
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": " get Item Image successfully ",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
...
    "/merchant/category/merchant": {
      "parameters": [
        {
          "$ref": "#/parameters/clientId"
        },
        {
          "$ref": "#/parameters/clientSecret"
        }
      ],
      "post": {
        "tags": [
          "MerchantCategory"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for add merchant category ",
        "consumes": [
          "application/json"
        ],
        "operationId": "addMerchantCategoriesByMerchant",
        "description": "API for add merchant category\n",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "category": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "update Order successfully ",
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "put": {
        "tags": [
          "MerchantCategory"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for update merchant category ",
        "consumes": [
          "application/json"
        ],
        "operationId": "updateMerchantCategoriesByMerchant",
        "description": "API for update merchant category\n",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "category": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "updatedRequest": {
                  "required": [
                    "isDeleted"
                  ],
                  "properties": {
                    "isDeleted": {
                      "description": "isDeleted false for Available, isDeleted true for not Available ",
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "update Order successfully ",
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "get": {
        "tags": [
          "MerchantCategory"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for getMerchantCategoriesByMerchant ",
        "consumes": [
          "application/json"
        ],
        "operationId": "getMerchantCategoriesByMerchant",
        "description": "API for get merchant category\n",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "minimum": 0,
            "required": false,
            "type": "number"
          },
          {
            "in": "query",
            "name": "limit",
            "minimum": 0,
            "maximum": 20,
            "required": false,
            "type": "number"
          },
          {
            "in": "query",
            "name": "categoryId",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "searchText",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "update Order successfully ",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/merchant/category/unselected/merchant": {
      "parameters": [
        {
          "$ref": "#/parameters/clientId"
        },
        {
          "$ref": "#/parameters/clientSecret"
        }
      ],
      "get": {
        "tags": [
          "MerchantCategory"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for get unselected merchant category ",
        "consumes": [
          "application/json"
        ],
        "operationId": "getUnselectedMerchantCategoriesByMerchant",
        "description": "API for get unselected merchant category\n",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "minimum": 0,
            "required": false,
            "type": "number"
          },
          {
            "in": "query",
            "name": "limit",
            "minimum": 0,
            "maximum": 20,
            "required": false,
            "type": "number"
          },
          {
            "in": "query",
            "name": "searchText",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "update Order successfully ",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/merchant/category/delete/merchant": {
      "parameters": [
        {
          "$ref": "#/parameters/clientId"
        },
        {
          "$ref": "#/parameters/clientSecret"
        }
      ],
      "put": {
        "tags": [
          "MerchantCategory"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for delete merchant category ",
        "consumes": [
          "application/json"
        ],
        "operationId": "deleteMerchantCategoriesByMerchant",
        "description": "API for delete merchant category\n",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "ids": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "update Order successfully ",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
...
  },
...
}